refactor: use class for recorder
