+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
-export * from './store/listener'
 export * from './store/node'
 export * from './store/idb'
 export * from './redux'
