--- packages/utils/src/store/listener.ts
@@ -1,1 +1,0 @@
-export const uninstallStore = new Set<Function>()
